optimizer/requirements.txt
torch
torchvision
loguru
numpy
pyzmq
